More code cleanups
