+++ internal/pkg/configuration/init.go
@@ -3,7 +3,7 @@ package configuration
 import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
-	"github.com/thepieterdc/gopos/pkg/logging"
+	"github.com/thepieterdc/gopos/internal/pkg/logging"
 )
 
 const (
