--- pkg/health/models.go
@@ -1,7 +1,0 @@
-package health
-
-// Response result of the /health route.
-type Response struct {
-	Status  bool   `json:"status"`
-	Version string `json:"version"`
-}
