+++ pkg/web/routes/google.go
@@ -3,8 +3,8 @@ package routes
 import (
 	"github.com/labstack/echo/v4"
 	log "github.com/sirupsen/logrus"
-	"github.com/thepieterdc/gopos/pkg/configuration"
-	"github.com/thepieterdc/gopos/pkg/logging"
+	"github.com/thepieterdc/gopos/internal/pkg/configuration"
+	"github.com/thepieterdc/gopos/internal/pkg/logging"
 	"github.com/thepieterdc/gopos/pkg/web/routes/google"
 	"net/http"
 )
