+++ pkg/web/routes/health.go
@@ -3,13 +3,18 @@ package routes
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/thepieterdc/gopos/internal/pkg/version"
-	"github.com/thepieterdc/gopos/pkg/health"
 	"net/http"
 )
 
+// healthResponse result of the /health route.
+type healthResponse struct {
+	Status  bool   `json:"status"`
+	Version string `json:"version"`
+}
+
 // HealthHandler handles the /health route.
 func HealthHandler(ctx echo.Context) error {
-	response := &health.Response{
+	response := &healthResponse{
 		Status:  true,
 		Version: version.VERSION,
 	}
